Cloud Computing Development/ml-flask-deployment/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
matplotlib